feat: extracted exit functions to a hook
